2020.12.17.0951 update

+++ src/main/java/com/example/demo_2020_1211/customer/Customer.java
@@ -2,6 +2,7 @@ package com.example.demo_2020_1211.customer;
 
 import javax.persistence.*;
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Entity //声明实体类
 @Table(name="cst_customer") //建立实体类和表的映射关系
@@ -10,7 +11,7 @@ public class Customer implements Serializable {
     @Id//声明当前私有属性为主键
     @GeneratedValue(strategy= GenerationType.IDENTITY) //配置主键的生成策略
     @Column(name="cust_id") //指定和表中cust_id字段的映射关系
-    private Long custId;
+    private BigDecimal custId;
 
     @Column(name="cust_name") //指定和表中cust_name字段的映射关系
     private String custName;
@@ -30,10 +31,10 @@ public class Customer implements Serializable {
     @Column(name="cust_phone")//指定和表中cust_phone字段的映射关系
     private String custPhone;
 
-    public Long getCustId() {
+    public BigDecimal getCustId() {
         return custId;
     }
-    public void setCustId(Long custId) {
+    public void setCustId(BigDecimal custId) {
         this.custId = custId;
     }
     public String getCustName() {
